+++ src/main/java/com/lne_archers/LNE_ArchersMod.java
@@ -14,7 +14,7 @@ public class LNE_ArchersMod implements ModInitializer {
 	public static final String MOD_ID = "lne_archers";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static ConfigManager<ItemConfig> itemConfig = new ConfigManager<ItemConfig>
-			("items", Default.itemConfig)
+			("items_v1", Default.itemConfig)
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)