+++ src/main/java/com/lne_archers/item/weapons/WitherSpear.java
@@ -1,11 +1,39 @@
 package com.lne_archers.item.weapons;
 
-import more_rpg_loot.item.weapons.WitherMeleeWeapon;
+import com.google.common.collect.Multimap;
+import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.annotation.Nullable;
+import more_rpg_loot.api.WeaponPassives;
+import more_rpg_loot.api.WeaponTooltips;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.block.BlockState;
+import net.minecraft.client.item.TooltipContext;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.EntityAttribute;
+import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.SwordItem;
 import net.minecraft.item.ToolMaterial;
+import net.minecraft.text.Text;
+import net.minecraft.world.World;
+import net.spell_engine.api.item.ConfigurableAttributes;
 
-public class WitherSpear extends WitherMeleeWeapon {
+import java.util.List;
+
+public class WitherSpear extends SwordItem implements ConfigurableAttributes {
+    private Multimap<EntityAttribute, EntityAttributeModifier> attributes;
+    public void setAttributes(Multimap<EntityAttribute, EntityAttributeModifier> attributes) {
+        this.attributes = attributes;
+    }
+
+    @Override
+    public Multimap<EntityAttribute, EntityAttributeModifier> getAttributeModifiers(EquipmentSlot slot) {
+        if (attributes == null) {
+            return super.getAttributeModifiers(slot);
+        }
+        return slot == EquipmentSlot.MAINHAND ? attributes : super.getAttributeModifiers(slot);
+    }
+
     public WitherSpear(ToolMaterial material, Settings settings) {
         this(material, 1,2.4F, settings);
     }
@@ -21,5 +49,23 @@ public class WitherSpear extends WitherMeleeWeapon {
     @Override
     public boolean isSuitableFor(BlockState state) {
         return false;
+    }
+    @Override
+    public boolean postHit(ItemStack stack, LivingEntity target, LivingEntity attacker) {
+        if (FabricLoader.getInstance().isModLoaded("loot_n_explore")) {
+            WeaponPassives.WitherMelee(stack, target, attacker);
+        }
+        stack.damage(1, attacker, (e)->{
+            e.sendEquipmentBreakStatus(EquipmentSlot.MAINHAND);
+        });
+        return true;
+    }
+
+    @Override
+    public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
+        if (FabricLoader.getInstance().isModLoaded("loot_n_explore")) {
+            WeaponTooltips.WitherMelee(stack, world, tooltip, context);
+        }
+        super.appendTooltip(stack, world, tooltip, context);
     }
 }